chore: 🏷️ Change fields from `int64` to `int`

+++ api/interfaces/interfaces.go
@@ -82,8 +82,8 @@ type Loomball struct {
 	Id                    primitive.ObjectID `json:"_id,omitempty"       bson:"_id,omitempty"`
 	Name                  string             `json:"name"      bson:"name"`
 	Serial                int                `json:"serial" bson:"serial"`
-	EffectiveUntil        int64              `json:"effective_until"      bson:"effective_until"`
-	DecayUntil            int64              `json:"decay_until"      bson:"decay_until"`
+	EffectiveUntil        int                `json:"effective_until"      bson:"effective_until"`
+	DecayUntil            int                `json:"decay_until"      bson:"decay_until"`
 	MinimumProbability    float64            `json:"minimum_probability"      bson:"minimum_probability"`
 	GymRewardChancePlayer float64            `json:"gym_reward_chance_player"      bson:"gym_reward_chance_player"`
 	GymRewardChanceOwner  float64            `json:"gym_reward_chance_owner"      bson:"gym_reward_chance_owner"`
